Feat#149 jwt 어노테이션 구현 및 컨트롤러 적용

+++ backend_set/src/main/java/org/funding/keyword/service/KeywordService.java
@@ -4,6 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.funding.keyword.dao.KeywordDAO;
 import org.funding.keyword.dto.KeywordRequestDTO;
 import org.funding.keyword.vo.KeywordVO;
+import org.funding.user.dao.MemberDAO;
+import org.funding.user.vo.MemberVO;
+import org.funding.user.vo.enumType.Role;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,6 +14,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class KeywordService {
+    private final MemberDAO memberDAO;
 
     private final KeywordDAO keywordDAO;
 
@@ -30,11 +34,25 @@ public class KeywordService {
         return keywordDAO.selectKeywordsByCategoryId(categoryId);
     }
 
-    public void addKeyword(KeywordRequestDTO requestDTO) {
+    public void addKeyword(KeywordRequestDTO requestDTO, Long userId) {
+        MemberVO member = memberDAO.findById(userId);
+        if (member == null) {
+            throw new RuntimeException("해당 멤버는 존재하지 않습니다.");
+        }
+        if (member.getRole() != Role.ROLE_ADMIN) {
+            throw new RuntimeException("해당 멤버는 관리자가 아닙니다.");
+        }
         keywordDAO.insertKeyword(requestDTO);
     }
 
-    public void deleteKeyword(String name) {
+    public void deleteKeyword(String name, Long userId) {
+        MemberVO member = memberDAO.findById(userId);
+        if (member == null) {
+            throw new RuntimeException("해당 멤버는 존재하지 않습니다.");
+        }
+        if (member.getRole() != Role.ROLE_ADMIN) {
+            throw new RuntimeException("해당 멤버는 관리자가 아닙니다.");
+        }
         keywordDAO.deleteKeyword(name);
     }
 }
